# specs/002-phase2-web-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo App API
  description: RESTful API for Todo Full-Stack Web Application (Phase II)
  version: 1.0.0
  contact:
    name: Todo App Team

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.example.com
    description: Production server (TBD)

security:
  - bearerAuth: []

paths:
  /api/{user_id}/tasks:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
        description: User ID (must match JWT token user_id)
    get:
      summary: List all tasks
      description: Retrieve all tasks belonging to the authenticated user. URL user_id must match JWT token user_id.
      operationId: listTasks
      tags:
        - Tasks
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create task
      description: Create a new task for the authenticated user. URL user_id must match JWT token user_id.
      operationId: createTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/{user_id}/tasks/{id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
        description: User ID (must match JWT token user_id)
    get:
      summary: Get task by ID
      description: Retrieve a specific task by ID. Only returns task if it belongs to the authenticated user. URL user_id must match JWT token user_id.
      operationId: getTask
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update task
      description: Update an existing task. Only allows update if task belongs to the authenticated user. URL user_id must match JWT token user_id.
      operationId: updateTask
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete task
      description: Delete a task. Only allows deletion if task belongs to the authenticated user. URL user_id must match JWT token user_id.
      operationId: deleteTask
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Task deleted successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/{user_id}/tasks/{id}/complete:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
        description: User ID (must match JWT token user_id)
    patch:
      summary: Toggle task completion
      description: Toggle the completion status of a task. Only allows toggle if task belongs to the authenticated user. URL user_id must match JWT token user_id.
      operationId: toggleTaskCompletion
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Task ID
      responses:
        '200':
          description: Task completion status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Better Auth

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Task ID
        user_id:
          type: string
          description: User ID (from JWT token)
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
        description:
          type: string
          nullable: true
          maxLength: 1000
          description: Task description
        completed:
          type: boolean
          description: Task completion status
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
        description:
          type: string
          nullable: true
          maxLength: 1000
          description: Task description

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
        description:
          type: string
          nullable: true
          maxLength: 1000
          description: Task description

    Error:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  loc:
                    type: array
                    items:
                      type: string
                  msg:
                    type: string
                  type:
                    type: string

  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Forbidden:
      description: Forbidden - URL user_id does not match JWT token user_id
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    NotFound:
      description: Not found - task not found or does not belong to authenticated user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'